+++ src/schema.d.ts
@@ -87,6 +87,10 @@ export interface APRenderRep {
          */
         tileHeight?: number;
         /**
+         * Only valid for the `stacking-offset` renderer. A number between 0 and 1 representing the percentage of a cell's space that should be used to offset each piece in the stack. A value of 1 will lead to no overlap. A value of 0 will stack all the pieces directly on top of each other.
+         */
+        stackOffset?: number;
+        /**
          * Sometimes a board needs shaded areas, lines showing ownership of board edges, things like that. This is how those are indicated. Not all features are available for all board styles.
          */
         markers?: (